refactor crud path variables
