+++ src/main/java/hexlet/code/controller/TaskController.java
@@ -7,8 +7,6 @@ import hexlet.code.model.Task;
 import hexlet.code.repository.TaskRepository;
 import hexlet.code.service.TaskService;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.AllArgsConstructor;
@@ -23,21 +21,17 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
 import javax.validation.Valid;
-import static hexlet.code.controller.TaskController.TASK_CONTROLLER_PATH;
 import static org.springframework.http.HttpStatus.CREATED;
 
 
 @AllArgsConstructor
 @RestController
-@RequestMapping("${base-url}" + TASK_CONTROLLER_PATH)
+@RequestMapping("/api" + "/tasks")
 public class TaskController {
     private TaskRepository taskRepository;
     private TaskService taskService;
 
-    public static final String TASK_CONTROLLER_PATH = "/tasks";
-    public static final String ID = "/{id}";
 
     private static final String ONLY_TASK_OWNER =
             "@taskRepository.findById(#id).get().getAuthor().getEmail() == authentication.getName()";
@@ -48,17 +42,14 @@ public class TaskController {
             @ApiResponse(responseCode = "200", description = "Task was found"),
             @ApiResponse(responseCode = "404", description = "Task with this id wasn`t found")
     })
-    @GetMapping(ID)
+    @GetMapping("/{id}")
     public Task getTask(@PathVariable long id) {
         return taskRepository.findById(id).get();
     }
 
     @Operation(summary = "Get all tasks if no filtration is set."
             + " Else Retrieves all the elements that match the conditions defined by the specified predicate ")
-    @ApiResponses(@ApiResponse(responseCode = "200", content =
-    @Content (schema =
-        @Schema (implementation = Task.class))
-    ))
+    @ApiResponses(@ApiResponse(responseCode = "200"))
     @GetMapping
     public Iterable<Task> getAllTasks(@QuerydslPredicate final Predicate predicate) {
         return predicate == null ? taskRepository.findAll() : taskRepository.findAll(predicate);
@@ -78,7 +69,7 @@ public class TaskController {
             @ApiResponse(responseCode = "404", description = "Task with this id wasn`t found")
     })
     @PreAuthorize(ONLY_TASK_OWNER)
-    @PutMapping(ID)
+    @PutMapping("/{id}")
     public Task updateTask(@RequestBody @Valid TaskDto taskDto, @PathVariable long id) {
         return taskService.updateTask(taskDto, id);
     }
@@ -89,7 +80,7 @@ public class TaskController {
             @ApiResponse(responseCode = "404", description = "Task with this id wasn`t found")
     })
     @PreAuthorize(ONLY_TASK_OWNER)
-    @DeleteMapping(ID)
+    @DeleteMapping("/{id}")
     public void deleteTask(@PathVariable long id) {
         taskRepository.deleteById(id);
     }
