+++ src/main/java/hexlet/code/controller/UserController.java
@@ -6,8 +6,6 @@ import hexlet.code.dto.UserDto;
 import hexlet.code.model.User;
 import hexlet.code.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.AllArgsConstructor;
@@ -21,26 +19,22 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
 import javax.validation.Valid;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
-
-import static hexlet.code.controller.UserController.USER_CONTROLLER_PATH;
 import static org.springframework.http.HttpStatus.CREATED;
 
 
 @AllArgsConstructor
 @RestController
-@RequestMapping("${base-url}" + USER_CONTROLLER_PATH)
+@RequestMapping("/api" + "/users")
 public class UserController {
 
+
     public static final String USER_CONTROLLER_PATH = "/users";
-    public static final String ID = "/{id}";
 
     private final UserService userService;
-
     private final UserRepository userRepository;
 
     private static final String ONLY_OWNER_BY_ID = """
@@ -53,16 +47,13 @@ public class UserController {
             @ApiResponse(responseCode = "200", description = "User was found"),
             @ApiResponse(responseCode = "404", description = "User with this id wasn`t found")
     })
-    @GetMapping(ID)
+    @GetMapping("/{id}")
     public Optional<User> getUser(@PathVariable long id) throws NoSuchElementException {
         return userRepository.findById(id);
     }
 
     @Operation(summary = "Get all users")
-    @ApiResponses(@ApiResponse(responseCode = "200", content =
-    @Content(schema =
-    @Schema(implementation = User.class))
-    ))
+    @ApiResponses(@ApiResponse(responseCode = "200"))
     @GetMapping
     public List<User> getAll() throws Exception {
         return userRepository.findAll()
@@ -84,7 +75,7 @@ public class UserController {
             @ApiResponse(responseCode = "404", description = "User with this id wasn`t found")
     })
     @PreAuthorize(ONLY_OWNER_BY_ID)
-    @PutMapping(ID)
+    @PutMapping("/{id}")
     public User updateUser(@PathVariable @Valid long id, @RequestBody @Valid UserDto userDto) {
         return userService.updateUser(id, userDto);
     }
@@ -95,7 +86,7 @@ public class UserController {
             @ApiResponse(responseCode = "404", description = "User with this id wasn`t found")
     })
     @PreAuthorize(ONLY_OWNER_BY_ID)
-    @DeleteMapping(ID)
+    @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable long id) {
         userRepository.deleteById(id);
     }
