+++ src/test/java/hexlet/code/controller/LabelConrollerIT.java
@@ -14,11 +14,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-
 import java.util.List;
 
-import static hexlet.code.utils.TestUtils.BASE_URL;
-import static hexlet.code.utils.TestUtils.TEST_USERNAME;
+import static hexlet.code.utils.TestUtils.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -42,8 +40,6 @@ class LabelConrollerIT {
     @Autowired
     private LabelRepository labelRepository;
 
-    public static final String LABEL_CONTROLLER_PATH = "/labels";
-    public static final String ID = "/{id}";
 
 
     @BeforeEach
