+++ src/test/java/hexlet/code/controller/TaskStatusControllerIT.java
@@ -16,9 +16,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import java.util.List;
-import static hexlet.code.controller.TaskStatusController.TASK_STATUS_CONTROLLER_PATH;
-import static hexlet.code.controller.UserController.ID;
 import static hexlet.code.utils.TestUtils.BASE_URL;
+import static hexlet.code.utils.TestUtils.ID;
+import static hexlet.code.utils.TestUtils.TASK_STATUS_CONTROLLER_PATH;
 import static hexlet.code.utils.TestUtils.TEST_STATUS_NAME;
 import static hexlet.code.utils.TestUtils.TEST_STATUS_NAME1;
 import static hexlet.code.utils.TestUtils.TEST_USERNAME;
